changed the force trees in the multi plane model to the quad trees generalized the slope for the pseudo NFW model now it can be any integer value > 0 then the slope is 0, the program exits
 changed FullRange/internal_rayshooter_multi.cpp
-- removed som printout statments

 changed MultiPlane/multiplane.cpp
-- changed to QuadTree

 changed TreeCode/forceTree.cpp
-- changed the slope beta to be int

 changed TreeCode/profiles.cpp
-- generalized the mass integral mhat for any integer value of the slope beta

 changed TreeCode/quadForceTrees.cpp
-- changed the slope beta to be int

 changed TreeCode/quad_profiles.cpp
-- generalized the mass integral mhat for any integer value of the slope beta

 changed include/forceTree.h
 changed include/multiplane.h
 changed include/quadTree.h
-- changed to QuadTree
-- changed the slope beta to be int

+++ include/quadTree.h
@@ -314,13 +314,13 @@ private:
 class QuadTreePseudoNFW : public QuadTree{
 
 public:
-	QuadTreePseudoNFW(float beta,PosType **xp,IndexType Npoints,HaloStructure *par_internals
+	QuadTreePseudoNFW(int beta,PosType **xp,IndexType Npoints,HaloStructure *par_internals
 			,double my_kappa_bk = 0.0,int bucket = 5,PosType theta = 0.1);
 	~QuadTreePseudoNFW();
 
 private:
 
-	double beta;
+	int beta;
 
 	// Override internal structure of halos
 	double alpha_h(double r2,HaloStructure &par);
